requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
matplotlib
joblib
librosa>=0.9.2
pydub
imagehash
Pillow
nltk
umap-learn
python-speech-features
SpeechRecognition
webrtcvad
transformers
torch
faiss-cpu ; platform_system != 'Windows'